Move persister into persistance package
