+++ persistance/persister.go
@@ -1,4 +1,4 @@
-package raft
+package persistance
 
 //
 // support for Raft and kvraft to save persistent
